Only allow kill process to be ran during a stop command

+++ docker/entrypoint.sh
@@ -22,6 +22,10 @@ build_module_simulator() {
 }
 
 kill_process() {
+  if [[ "$COMMAND" != "stop" ]]; then
+    echo "Cannot run kill_process outside of the stop command."
+    exit 1
+  fi
   pkill -9 -f $1
 }
